refactor(lsp): extract download progress listener into its own component
Move the throttling progress logic out of the provider into a dedicated
ProgressIndicatorDownloadListener (implements DownloadProgressListener),
instead of exposing a package-private factory method as a test seam. The
component is unit-tested directly in ProgressIndicatorDownloadListenerTest;
the provider test keeps mocking the downloader, so the listener is not
exercised there.
