+++ src/test/java/com/github/_1c_syntax/bsl/intellij/lsp/ProgressIndicatorDownloadListenerTest.java
@@ -1,0 +1,90 @@
+/*
+ * This file is a part of IntelliJ Language 1C (BSL) Plugin.
+ *
+ * Copyright © 2018-2026
+ * Alexey Sosnoviy <labotamy@gmail.com>, Nikita Fedkin <nixel2007@gmail.com>
+ *
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * IntelliJ Language 1C (BSL) Plugin is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * IntelliJ Language 1C (BSL) Plugin is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with IntelliJ Language 1C (BSL) Plugin.
+ */
+package com.github._1c_syntax.bsl.intellij.lsp;
+
+import com.intellij.openapi.progress.ProcessCanceledException;
+import com.intellij.openapi.progress.ProgressIndicator;
+import com.intellij.testFramework.fixtures.BasePlatformTestCase;
+
+import static org.mockito.ArgumentMatchers.anyDouble;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+/**
+ * Тесты отображения прогресса скачивания на индикаторе IntelliJ: троттлинг перерисовки, доля/текст
+ * при известном размере, неопределённый режим при неизвестном и проброс отмены задачи.
+ */
+public class ProgressIndicatorDownloadListenerTest extends BasePlatformTestCase {
+
+  public void testReportsFractionAndThrottlesWithinPercent() {
+    var total = 200L * 1024 * 1024; // 200 МБ, 1% == 2 МБ
+    var indicator = mock(ProgressIndicator.class);
+    var listener = new ProgressIndicatorDownloadListener(indicator);
+
+    listener.onProgress(0, total);                 // 0% — первая перерисовка
+    listener.onProgress(1024L * 1024, total);      // всё ещё 0% — троттлится
+    listener.onProgress(2L * 1024 * 1024, total);  // 1% — новая перерисовка
+
+    // Три уведомления, но индикатор перерисован только на смене процента — дважды.
+    verify(indicator, times(2)).setFraction(anyDouble());
+    verify(indicator).setFraction(0.0);
+    // setIndeterminate(false) выставляется на каждой перерисовке — их две.
+    verify(indicator, times(2)).setIndeterminate(false);
+    verify(indicator).setText2("0.0 MB / 200.0 MB");
+    verify(indicator).setText2("2.0 MB / 200.0 MB");
+    // Отмену проверяем на каждом блоке, даже на пропущенном троттлингом.
+    verify(indicator, times(3)).checkCanceled();
+  }
+
+  public void testIsIndeterminateWhenTotalUnknown() {
+    var indicator = mock(ProgressIndicator.class);
+    var listener = new ProgressIndicatorDownloadListener(indicator);
+
+    listener.onProgress(512L * 1024, -1);   // 0 МБ — первая перерисовка
+    listener.onProgress(768L * 1024, -1);   // всё ещё 0 МБ — троттлится
+    listener.onProgress(1024L * 1024, -1);  // 1 МБ — новая перерисовка
+
+    // Размер неизвестен: доля не выставляется, индикатор — неопределённый.
+    verify(indicator, never()).setFraction(anyDouble());
+    verify(indicator, times(2)).setIndeterminate(true);
+    verify(indicator).setText2("0.5 MB");
+    verify(indicator).setText2("1.0 MB");
+    verify(indicator, times(3)).checkCanceled();
+  }
+
+  public void testPropagatesCancellation() {
+    var indicator = mock(ProgressIndicator.class);
+    doThrow(new ProcessCanceledException()).when(indicator).checkCanceled();
+    var listener = new ProgressIndicatorDownloadListener(indicator);
+
+    try {
+      listener.onProgress(1024, 4096);
+      fail("Expected ProcessCanceledException to abort the download");
+    } catch (ProcessCanceledException expected) {
+      // отмена индикатора должна прерывать загрузку
+    }
+    verify(indicator, never()).setFraction(anyDouble());
+  }
+}
